refactor: extract MCP OAuth core module
- move OAuth discovery, callback, PKCE, token exchange, and refresh helpers into src/core/oauth.rs
- keep CLI OAuth add/remove command flow but route through shared core OAuth utilities
- switch MCP client refresh dependency from crate::cli to crate::core::oauth
- register the new core oauth module and move OAuth helper tests into it
- update README architecture overview to document the shared OAuth module
